Initial Commit
I've got a GUI program that stores an incoming clipboard state, then displays a series of buttons, each of which puts things in the clipboard.

Each button represents a field it found in the pass entry.

Finally, it restores the previous clipboard when the program closes.

It's a little tight on space right now, but it's otherwise pretty workable.
The one downside is that it's all hard-coded to one pass entry.
I'll have to do fuzzy search next.