Added Launcher
I was going to build my own fuzzy-finder thing.
But then I was like "Well, I _have_ dmenu"
So now I use that.

So launch is a very simple script that lists the entries and gives them to dmenu.
Then it takes what dmenu selects and gives _that_ to passtk, which has been extended to allow selection of items from a commandline arg.

At this point it _basically_ works...