Pass OTP Support
This is a real hack, but pass-otp dumps something in the file that looks like:
otpauth://totp/blahblahblah

And I'm already splitting on colon...
So I would get a label with the name "otpauth" and the contents of //totp/stuff

So now instead I just recognize that label name, and when I see it I just instead run
  pass otp NAME
instead and copy that to the clipboard rather than whatever I have.

Which is great!
Simple, and nothing could ever go wrong!